+++ src/MicroOcpp_c.h
@@ -9,7 +9,6 @@
 
 #include <MicroOcpp/Core/ConfigurationOptions.h>
 #include <MicroOcpp/Model/ConnectorBase/ChargePointStatus.h>
-#include <MicroOcpp/Model/ConnectorBase/Notification.h>
 #include <MicroOcpp/Model/ConnectorBase/UnlockConnectorResult.h>
 #include <MicroOcpp/Model/Transactions/Transaction.h>
 #include <MicroOcpp/Model/Certificates/Certificate_c.h>
@@ -61,14 +60,16 @@ void ocpp_initialize(
             const char *chargePointModel,     //model name of this charger (e.g. "My Charger")
             const char *chargePointVendor, //brand name (e.g. "My Company Ltd.")
             struct OCPP_FilesystemOpt fsopt, //If this library should format the flash if necessary. Find further options in ConfigurationOptions.h
-            bool autoRecover); //automatically sanitize the local data store when the lib detects recurring crashes. During development, `false` is recommended
+            bool autoRecover, //automatically sanitize the local data store when the lib detects recurring crashes. During development, `false` is recommended
+            bool ocpp201); //true to select OCPP 2.0.1, false for OCPP 1.6
 
 //same as above, but more fields for the BootNotification
 void ocpp_initialize_full(
             OCPP_Connection *conn,  //WebSocket adapter for MicroOcpp
             const char *bootNotificationCredentials, //e.g. '{"chargePointModel":"Demo Charger","chargePointVendor":"My Company Ltd."}' (refer to OCPP 1.6 Specification - Edition 2 p. 60)
             struct OCPP_FilesystemOpt fsopt, //If this library should format the flash if necessary. Find further options in ConfigurationOptions.h
-            bool autoRecover); //automatically sanitize the local data store when the lib detects recurring crashes. During development, `false` is recommended
+            bool autoRecover, //automatically sanitize the local data store when the lib detects recurring crashes. During development, `false` is recommended
+            bool ocpp201); //true to select OCPP 2.0.1, false for OCPP 1.6
 
 void ocpp_deinitialize();
 
@@ -78,11 +79,11 @@ void ocpp_loop();
  * Charging session management
  */
 
-void ocpp_beginTransaction(const char *idTag);
-void ocpp_beginTransaction_m(unsigned int connectorId, const char *idTag); //multiple connectors version
+bool ocpp_beginTransaction(const char *idTag);
+bool ocpp_beginTransaction_m(unsigned int connectorId, const char *idTag); //multiple connectors version
 
-void ocpp_beginTransaction_authorized(const char *idTag, const char *parentIdTag);
-void ocpp_beginTransaction_authorized_m(unsigned int connectorId, const char *idTag, const char *parentIdTag);
+bool ocpp_beginTransaction_authorized(const char *idTag, const char *parentIdTag);
+bool ocpp_beginTransaction_authorized_m(unsigned int connectorId, const char *idTag, const char *parentIdTag);
 
 bool ocpp_endTransaction(const char *idTag, const char *reason); //idTag, reason can be NULL
 bool ocpp_endTransaction_m(unsigned int connectorId, const char *idTag, const char *reason); //idTag, reason can be NULL
@@ -156,8 +157,8 @@ void ocpp_setStartTxReadyInput_m(unsigned int connectorId, InputBool_m startTxRe
 void ocpp_setStopTxReadyInput(InputBool stopTxReady);
 void ocpp_setStopTxReadyInput_m(unsigned int connectorId, InputBool_m stopTxReady);
 
-void ocpp_setTxNotificationOutput(void (*notificationOutput)(OCPP_Transaction*, enum OCPP_TxNotification));
-void ocpp_setTxNotificationOutput_m(unsigned int connectorId, void (*notificationOutput)(unsigned int, OCPP_Transaction*, enum OCPP_TxNotification));
+void ocpp_setTxNotificationOutput(void (*notificationOutput)(OCPP_Transaction*, TxNotification));
+void ocpp_setTxNotificationOutput_m(unsigned int connectorId, void (*notificationOutput)(unsigned int, OCPP_Transaction*, TxNotification));
 
 #if MO_ENABLE_CONNECTOR_LOCK
 void ocpp_setOnUnlockConnectorInOut(PollUnlockResult onUnlockConnectorInOut);