+++ tests/benchmarks/scripts/eval_firmware_size.py
@@ -100,10 +100,17 @@ def categorize_table(df):
     df.at['Core/Time.cpp', 'v16'] = TICK
     df.at['Core/Time.cpp', 'v201'] = TICK
     df.at['Core/Time.cpp', 'Module'] = MODULE_GENERAL
+    df.at['Core/UuidUtils.cpp', 'v16'] = TICK
+    df.at['Core/UuidUtils.cpp', 'v201'] = TICK
+    df.at['Core/UuidUtils.cpp', 'Module'] = MODULE_GENERAL
     if 'Debug.cpp' in df.index:
         df.at['Debug.cpp', 'v16'] = TICK
         df.at['Debug.cpp', 'v201'] = TICK
         df.at['Debug.cpp', 'Module'] = MODULE_HAL
+    if 'Platform.cpp' in df.index:
+        df.at['Platform.cpp', 'v16'] = TICK
+        df.at['Platform.cpp', 'v201'] = TICK
+        df.at['Platform.cpp', 'Module'] = MODULE_HAL
     df.at['Model/Authorization/AuthorizationData.cpp', 'v16'] = TICK
     df.at['Model/Authorization/AuthorizationData.cpp', 'Module'] = MODULE_LOCALAUTH
     df.at['Model/Authorization/AuthorizationList.cpp', 'v16'] = TICK
@@ -137,9 +144,6 @@ def categorize_table(df):
     df.at['Model/ConnectorBase/Connector.cpp', 'Module'] = MODULE_CORE
     df.at['Model/ConnectorBase/ConnectorsCommon.cpp', 'v16'] = TICK
     df.at['Model/ConnectorBase/ConnectorsCommon.cpp', 'Module'] = MODULE_CORE
-    if 'Model/ConnectorBase/Notification.cpp' in df.index:
-        df.at['Model/ConnectorBase/Notification.cpp', 'v16'] = TICK
-        df.at['Model/ConnectorBase/Notification.cpp', 'Module'] = MODULE_CORE
     df.at['Model/Diagnostics/DiagnosticsService.cpp', 'v16'] = TICK
     df.at['Model/Diagnostics/DiagnosticsService.cpp', 'Module'] = MODULE_FW_MNGT
     df.at['Model/FirmwareManagement/FirmwareService.cpp', 'v16'] = TICK